requirements.txt
streamlit
langchain
google-generativeai
openai
langchain-google-genai

